Move utils/fitbm_problem.py to parsing
