+++ fitbenchmarking/parsing/parse_fitbenchmark_data.py
@@ -29,7 +29,6 @@ import os
 import re
 import numpy as np
 
-from utils import fitbm_problem
 from fitting.mantid.externals import store_main_problem_data
 from utils.logging_setup import logger
